- 로그인 시 상태프로그래스 바 추가 - 로그인 페이지 구조 개선
