+++ lib/components/progress_bar.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class ProgressBar extends StatelessWidget {
+  const ProgressBar({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: CircularProgressIndicator(),
+    );
+  }
+}
