+++ lib/state/login.dart
@@ -1,0 +1,5 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+final loginProgressBarProvider = StateProvider<bool>((ref) {
+  return false;
+});
